consolidate bot eligibility onto the model
Move BotEligibilityChecker into the Tournaments namespace and have it
return a Result value object instead of a bare hash, so the result
shape is owned in one place. Add Bot#eligibility_for / Bot#eligible_for?
and migrate all three controllers to them, deleting the BotEligibility
controller concern. Namespace the constraints-params concern to
Tournaments::ConstraintsParams to match.

Drop the redundant client-side eligibility filter: the server already
renders only eligible bots, so the eligibility-check Stimulus
controller, the eligible_bots endpoint/route, and that JS are removed.
@eligible_bots is now a stable array built through the compiled scope.

Adds Bot eligibility specs and a deterministic open-registration spec.
