--- app/controllers/concerns/bot_eligibility.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module BotEligibility
-  private
-
-  def check_bot_eligibility(bot, constraints)
-    BotEligibilityChecker.new(bot.compiled_program, constraints).check
-  end
-end
