+++ app/controllers/tournament_entries_controller.rb
@@ -1,5 +1,4 @@
 class TournamentEntriesController < ApplicationController
-  include BotEligibility
   before_action :authenticate_registered_user!
   before_action :set_tournament
 
@@ -147,9 +146,9 @@ class TournamentEntriesController < ApplicationController
 
   def constraint_violation_message(bot)
     return nil unless tournament.constraints.present?
-    result = check_bot_eligibility(bot, tournament.constraints)
-    return nil if result[:eligible]
-    "Bot is not eligible for this tournament: #{result[:violations].map { |v| v[:message] }.join(', ')}"
+    result = bot.eligibility_for(tournament.constraints)
+    return nil if result.eligible?
+    "Bot is not eligible for this tournament: #{result.violations.map { |v| v[:message] }.join(', ')}"
   end
 
   def entry_params
