+++ spec/services/tournaments/bot_eligibility_checker_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe BotEligibilityChecker do
+RSpec.describe Tournaments::BotEligibilityChecker do
   def program(*node_defs)
     nodes = node_defs.each_with_object({}) do |n, h|
       h[n[:id]] = n.stringify_keys.merge("children" => n.fetch(:children, []))
@@ -46,19 +46,19 @@ RSpec.describe BotEligibilityChecker do
   describe "with nil constraints" do
     it "is always eligible" do
       result = check(linear_program, nil)
-      expect(result[:eligible]).to be true
-      expect(result[:violations]).to be_empty
+      expect(result.eligible).to be true
+      expect(result.violations).to be_empty
     end
 
     it "returns cost 0 and nil budget" do
       result = check(linear_program, nil)
-      expect(result[:cost]).to eq(0)
-      expect(result[:budget]).to be_nil
+      expect(result.cost).to eq(0)
+      expect(result.budget).to be_nil
     end
 
     it "populates all stats keys" do
       result = check(linear_program, nil)
-      expect(result[:stats].keys).to contain_exactly(:and_count, :or_count, :score_node_count, :max_branch_length)
+      expect(result.stats.keys).to contain_exactly(:and_count, :or_count, :score_node_count, :max_branch_length)
     end
   end
 
@@ -75,7 +75,7 @@ RSpec.describe BotEligibilityChecker do
     end
 
     it "counts a node with 2 parents as 1 OR" do
-      expect(check(or_program)[:stats][:or_count]).to eq(1)
+      expect(check(or_program).stats[:or_count]).to eq(1)
     end
 
     # root → c1 → s1
@@ -89,7 +89,7 @@ RSpec.describe BotEligibilityChecker do
         condition("c3", children: ["s1"]),
         score("s1")
       )
-      expect(check(prog)[:stats][:or_count]).to eq(2)
+      expect(check(prog).stats[:or_count]).to eq(2)
     end
 
     it "does not count a fork that never reconverges as an OR" do
@@ -102,7 +102,7 @@ RSpec.describe BotEligibilityChecker do
         score("s1"),
         score("s2")
       )
-      expect(check(prog)[:stats][:or_count]).to eq(0)
+      expect(check(prog).stats[:or_count]).to eq(0)
     end
 
     it "counts an OR through an organizer" do
@@ -114,7 +114,7 @@ RSpec.describe BotEligibilityChecker do
         organizer("o2", children: ["s1"]),
         score("s1")
       )
-      expect(check(prog)[:stats][:or_count]).to eq(1)
+      expect(check(prog).stats[:or_count]).to eq(1)
     end
 
     context "when allow_or is false" do
@@ -122,12 +122,12 @@ RSpec.describe BotEligibilityChecker do
 
       it "adds a violation when ORs are present" do
         result = check(or_program, constraints)
-        expect(result[:eligible]).to be false
-        expect(result[:violations]).to include(include(type: "allow_or"))
+        expect(result.eligible).to be false
+        expect(result.violations).to include(include(type: "allow_or"))
       end
 
       it "is eligible when no ORs are present" do
-        expect(check(linear_program, constraints)[:eligible]).to be true
+        expect(check(linear_program, constraints).eligible).to be true
       end
     end
   end
@@ -144,7 +144,7 @@ RSpec.describe BotEligibilityChecker do
     end
 
     it "counts one condition child as 1 AND" do
-      expect(check(and_program)[:stats][:and_count]).to eq(1)
+      expect(check(and_program).stats[:and_count]).to eq(1)
     end
 
     it "counts two condition children as 2 ANDs" do
@@ -157,7 +157,7 @@ RSpec.describe BotEligibilityChecker do
       )
       # c1 has 2 condition children; c2 and c3 have no condition children
       # s1 has 2 parents → 1 OR (separate from AND count)
-      expect(check(prog)[:stats][:and_count]).to eq(2)
+      expect(check(prog).stats[:and_count]).to eq(2)
     end
 
     it "does not count an organizer child as an AND" do
@@ -168,11 +168,11 @@ RSpec.describe BotEligibilityChecker do
         condition("c2", children: ["s1"]),
         score("s1")
       )
-      expect(check(prog)[:stats][:and_count]).to eq(0)
+      expect(check(prog).stats[:and_count]).to eq(0)
     end
 
     it "does not count a score child as an AND" do
-      expect(check(linear_program)[:stats][:and_count]).to eq(0)
+      expect(check(linear_program).stats[:and_count]).to eq(0)
     end
 
     context "when allow_and is false" do
@@ -180,12 +180,12 @@ RSpec.describe BotEligibilityChecker do
 
       it "adds a violation when ANDs are present" do
         result = check(and_program, constraints)
-        expect(result[:eligible]).to be false
-        expect(result[:violations]).to include(include(type: "allow_and"))
+        expect(result.eligible).to be false
+        expect(result.violations).to include(include(type: "allow_and"))
       end
 
       it "is eligible when no ANDs are present" do
-        expect(check(linear_program, constraints)[:eligible]).to be true
+        expect(check(linear_program, constraints).eligible).to be true
       end
     end
   end
@@ -201,7 +201,7 @@ RSpec.describe BotEligibilityChecker do
         condition("c2", children: ["s1"]),
         score("s1")
       )
-      expect(check(prog)[:stats][:max_branch_length]).to eq(2)
+      expect(check(prog).stats[:max_branch_length]).to eq(2)
     end
 
     it "returns the longest path when branches diverge" do
@@ -215,12 +215,12 @@ RSpec.describe BotEligibilityChecker do
         score("s1"),
         score("s2")
       )
-      expect(check(prog)[:stats][:max_branch_length]).to eq(2)
+      expect(check(prog).stats[:max_branch_length]).to eq(2)
     end
 
     it "is 0 for a program with no condition nodes" do
       prog = program(root("r", children: ["s1"]), score("s1"))
-      expect(check(prog)[:stats][:max_branch_length]).to eq(0)
+      expect(check(prog).stats[:max_branch_length]).to eq(0)
     end
 
     context "with max_branch_length constraint" do
@@ -234,8 +234,8 @@ RSpec.describe BotEligibilityChecker do
           score("s1")
         )
         result = check(prog, { "max_branch_length" => 2 })
-        expect(result[:eligible]).to be false
-        expect(result[:violations]).to include(include(type: "max_branch_length"))
+        expect(result.eligible).to be false
+        expect(result.violations).to include(include(type: "max_branch_length"))
       end
 
       it "is eligible when exactly at the limit" do
@@ -245,7 +245,7 @@ RSpec.describe BotEligibilityChecker do
           condition("c2", children: ["s1"]),
           score("s1")
         )
-        expect(check(prog, { "max_branch_length" => 2 })[:eligible]).to be true
+        expect(check(prog, { "max_branch_length" => 2 }).eligible).to be true
       end
     end
   end
@@ -259,7 +259,7 @@ RSpec.describe BotEligibilityChecker do
         score("s1"),
         score("s2")
       )
-      expect(check(prog)[:stats][:score_node_count]).to eq(2)
+      expect(check(prog).stats[:score_node_count]).to eq(2)
     end
 
     context "with max_score_nodes constraint" do
@@ -272,12 +272,12 @@ RSpec.describe BotEligibilityChecker do
           score("s2")
         )
         result = check(prog, { "max_score_nodes" => 1 })
-        expect(result[:eligible]).to be false
-        expect(result[:violations]).to include(include(type: "max_score_nodes"))
+        expect(result.eligible).to be false
+        expect(result.violations).to include(include(type: "max_score_nodes"))
       end
 
       it "is eligible when exactly at the limit" do
-        expect(check(linear_program, { "max_score_nodes" => 1 })[:eligible]).to be true
+        expect(check(linear_program, { "max_score_nodes" => 1 }).eligible).to be true
       end
     end
   end
@@ -296,7 +296,7 @@ RSpec.describe BotEligibilityChecker do
     it "sums costs by node type and kind" do
       # 1 unary condition (3) + 1 score node (2) = 5
       result = check(linear_program, { "costs" => costs })
-      expect(result[:cost]).to eq(5)
+      expect(result.cost).to eq(5)
     end
 
     it "charges and_cost per AND edge" do
@@ -307,7 +307,7 @@ RSpec.describe BotEligibilityChecker do
         condition("c2", children: ["s1"]),
         score("s1")
       )
-      expect(check(prog, { "costs" => costs })[:cost]).to eq(9)
+      expect(check(prog, { "costs" => costs }).cost).to eq(9)
     end
 
     it "charges or_cost per OR" do
@@ -318,7 +318,7 @@ RSpec.describe BotEligibilityChecker do
         condition("c2", children: ["s1"]),
         score("s1")
       )
-      expect(check(prog, { "costs" => costs })[:cost]).to eq(12)
+      expect(check(prog, { "costs" => costs }).cost).to eq(12)
     end
 
     it "uses the kind-specific cost for relational conditions" do
@@ -328,24 +328,24 @@ RSpec.describe BotEligibilityChecker do
         score("s1")
       )
       # relational (5) + score (2) = 7
-      expect(check(prog, { "costs" => costs })[:cost]).to eq(7)
+      expect(check(prog, { "costs" => costs }).cost).to eq(7)
     end
 
     it "skips cost keys that are nil" do
       # Only score_node cost set; condition has no cost
       result = check(linear_program, { "costs" => { "score_node" => 10 } })
-      expect(result[:cost]).to eq(10)
+      expect(result.cost).to eq(10)
     end
 
     it "adds a budget violation when cost exceeds budget" do
       result = check(linear_program, { "costs" => costs, "budget" => 4 })
-      expect(result[:eligible]).to be false
-      expect(result[:violations]).to include(include(type: "budget"))
+      expect(result.eligible).to be false
+      expect(result.violations).to include(include(type: "budget"))
     end
 
     it "is eligible when cost equals budget" do
       # 1 unary (3) + 1 score (2) = 5
-      expect(check(linear_program, { "costs" => costs, "budget" => 5 })[:eligible]).to be true
+      expect(check(linear_program, { "costs" => costs, "budget" => 5 }).eligible).to be true
     end
   end
 
@@ -353,7 +353,7 @@ RSpec.describe BotEligibilityChecker do
     context "with banned_action_types" do
       it "is eligible when no score nodes use a banned type" do
         result = check(linear_program, { "score_node_restrictions" => { "banned_action_types" => ["return"] } })
-        expect(result[:eligible]).to be true
+        expect(result.eligible).to be true
       end
 
       it "aggregates one violation per banned action type with a node count" do
@@ -365,14 +365,14 @@ RSpec.describe BotEligibilityChecker do
           score("s2", action_type: "return")
         )
         result = check(prog, { "score_node_restrictions" => { "banned_action_types" => ["return"] } })
-        violations = result[:violations].select { |v| v[:type] == "score_node_action_type" }
+        violations = result.violations.select { |v| v[:type] == "score_node_action_type" }
         expect(violations.size).to eq(1)
         expect(violations.first[:message]).to include("return").and include("2 nodes")
       end
 
       it "is eligible with no banned types set" do
         result = check(linear_program, { "score_node_restrictions" => {} })
-        expect(result[:eligible]).to be true
+        expect(result.eligible).to be true
       end
     end
   end
@@ -386,17 +386,17 @@ RSpec.describe BotEligibilityChecker do
           score("s1")
         )
         result = check(prog, { "condition_restrictions" => { "banned_kinds" => ["relational"] } })
-        expect(result[:violations]).to include(include(type: "condition_kind"))
+        expect(result.violations).to include(include(type: "condition_kind"))
       end
 
       it "is eligible when no conditions use a banned kind" do
         result = check(linear_program, { "condition_restrictions" => { "banned_kinds" => ["relational"] } })
-        expect(result[:eligible]).to be true
+        expect(result.eligible).to be true
       end
 
       it "is eligible with no banned kinds set" do
         result = check(linear_program, { "condition_restrictions" => {} })
-        expect(result[:eligible]).to be true
+        expect(result.eligible).to be true
       end
     end
 
@@ -412,18 +412,18 @@ RSpec.describe BotEligibilityChecker do
       it "adds a violation for a disallowed subject" do
         constraints = { "condition_restrictions" => { "unary" => { "allowed_subjects" => ["last_moved_piece"] } } }
         result = check(prog, constraints)
-        expect(result[:violations]).to include(include(type: "condition_subject"))
+        expect(result.violations).to include(include(type: "condition_subject"))
       end
 
       it "adds a violation for a disallowed operator" do
         constraints = { "condition_restrictions" => { "unary" => { "allowed_operators" => ["control"] } } }
         result = check(prog, constraints)
-        expect(result[:violations]).to include(include(type: "condition_operator"))
+        expect(result.violations).to include(include(type: "condition_operator"))
       end
 
       it "is eligible when all fields are within allowed lists" do
         constraints = { "condition_restrictions" => { "unary" => { "allowed_subjects" => ["moved_piece"], "allowed_operators" => ["value"] } } }
-        expect(check(prog, constraints)[:eligible]).to be true
+        expect(check(prog, constraints).eligible).to be true
       end
 
       it "does not apply kind-specific rules to conditions of a different kind" do
@@ -434,7 +434,7 @@ RSpec.describe BotEligibilityChecker do
           score("s1")
         )
         constraints = { "condition_restrictions" => { "unary" => { "allowed_subjects" => ["last_moved_piece"] } } }
-        expect(check(prog2, constraints)[:eligible]).to be true
+        expect(check(prog2, constraints).eligible).to be true
       end
     end
   end
@@ -451,8 +451,8 @@ RSpec.describe BotEligibilityChecker do
         "score_node_restrictions" => { "banned_action_types" => ["return"] }
       }
       result = check(prog, constraints)
-      expect(result[:eligible]).to be false
-      expect(result[:violations].map { |v| v[:type] }).to contain_exactly("condition_kind", "score_node_action_type")
+      expect(result.eligible).to be false
+      expect(result.violations.map { |v| v[:type] }).to contain_exactly("condition_kind", "score_node_action_type")
     end
   end
 end
